# dev/week8/goals-api/backend/Swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Workouts API",
      "contact": {},
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:4000/api",
        "variables": {}
      }
    ],
    "paths": {
      "/user/signup": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Register User",
          "operationId": "RegisterUser",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterUserRequest"
                },
                "example": {
                  "name": "bob",
                  "email": "bob@gmail.com",
                  "password": "4wa95#Cf-"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": []
        }
      },
      "/user/login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Login User",
          "operationId": "LoginUser",
          "parameters": [],
          "requestBody": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserRequest"
                },
                "example": {
                  "email": "susan@gmail.com",
                  "password": "secret"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "deprecated": false,
          "security": []
        }
      },
...
    },
    "components": {
      "schemas": {
        "RegisterUserRequest": {
          "title": "RegisterUserRequest",
          "required": [
            "name",
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "name": "bob",
            "email": "bob@gmail.com",
            "password": "4wa95#Cf-"
          }
        },
        "LoginUserRequest": {
          "title": "LoginUserRequest",
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "example": {
            "email": "bob@gmail.com",
            "password": "4wa95#Cf-"
          }
        },
...
      },
      "securitySchemes": {
        "httpBearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
    "security": [
      {
        "httpBearer": []
      }
    ],
...
  }